Fix result_score IndexError and correct score
The result_score parse was looking for a class that no longer exists so
I updated it to the current one. Similarly, the index of the scores
themselves appeared to be wrong, so I've also updated them. To prevent
IndexErrors breaking the program in the future, I've wrapped each of the
parsing statements in try/except blocks.